# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Rezervacija resursa",
        "description": "API za login, registraciju i upravljanje korisnicima",
        "termsOfService": "",
        "contact": {
            "email": "mislav.zabcic@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost",
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
        "items/details": {
            "get": {
                "tags": [
                    "details"
                ],
                "summary": "Dohvati sve detalje itema",
                "description": "Dohvati sve detalje itema",
                "operationId": "api.items.details",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Pisati u formatu <ime_kolone_u_tablici>=<pojam_za_pretra\u017eivanje>",
                        "required": true,
                        "type": "string",
                        "items": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalji itema",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/DetailsResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "501": {
                        "description": "Invalid search data",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "401": {
                        "description": "Token invalid",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "402": {
                        "description": "No token recived",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "410": {
                        "description": "Token expired",
                        "schema": {
                            "$ref": "#/definitions/TokenExpired"
                        }
                    }
                },
                "schemes": [
                    "http"
                ]
            }
        },
...
        "items/details/subtypes": {
            "get": {
                "tags": [
                    "details"
                ],
                "summary": "Dohvati sve podtipove ure\u0111aja ili filtrirane",
                "description": "Dohvati sve podtipove ure\u0111aja ili filtrirane",
                "operationId": "api.items.details.subtypes",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Pisati u formatu <ime_kolone_u_tablici>=<pojam_za_pretra\u017eivanje>",
                        "required": true,
                        "type": "string",
                        "items": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subtypes",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/SubType"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "501": {
                        "description": "Invalid search data",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "401": {
                        "description": "Token invalid",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "402": {
                        "description": "No token recived",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "410": {
                        "description": "Token expired",
                        "schema": {
                            "$ref": "#/definitions/TokenExpired"
                        }
                    }
                },
                "schemes": [
                    "http"
                ]
            }
        },
...
        "admin/items/details/create/devicetype": {
            "post": {
                "tags": [
                    "admin"
                ],
                "summary": "Kreiranje device tipa",
                "description": "Kreiranje device tipa",
                "operationId": "api.admin.items.details.create.devicetype",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "JWT token",
                        "required": true,
                        "type": "string",
                        "items": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "label",
                        "in": "body",
                        "description": "Labela",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "body",
                        "description": "Opis",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Device type created"
                    },
                    "409": {
                        "description": "Item with that label already exists",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "400": {
                        "description": "Invalid form data",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "401": {
                        "description": "Token invalid",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "402": {
                        "description": "No token recived",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "410": {
                        "description": "Token expired",
                        "schema": {
                            "$ref": "#/definitions/TokenExpired"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "403": {
                        "description": "No admin rights",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    }
                },
                "schemes": [
                    "http"
                ]
            }
        },
...
        "admin/items/details/create/subtype": {
            "post": {
                "tags": [
                    "admin"
                ],
                "summary": "Kreiranje  podtipa",
                "description": "Kreiranje podtipa",
                "operationId": "api.admin.items.details.create.subtype",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "JWT token",
                        "required": true,
                        "type": "string",
                        "items": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "label",
                        "in": "body",
                        "description": "Labela",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "body",
                        "description": "Opis",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subtype created"
                    },
                    "409": {
                        "description": "Item with that label already exists",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "400": {
                        "description": "Invalid form data",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "401": {
                        "description": "Token invalid",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "402": {
                        "description": "No token recived",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "410": {
                        "description": "Token expired",
                        "schema": {
                            "$ref": "#/definitions/TokenExpired"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "403": {
                        "description": "No admin rights",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    }
                },
                "schemes": [
                    "http"
                ]
            }
        },
...
    },
    "definitions": {
        "CustomError": {
            "properties": {
                "error": {
                    "description": "",
                    "type": "string",
                    "format": "string"
                }
            },
            "type": "object"
        },
        "DetailsResponse": {
            "properties": {
                "kits": {
                    "description": "",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Kit"
                    }
                },
                "types": {
                    "description": "",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Type"
                    }
                },
                "subtypes": {
                    "description": "",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/SubType"
                    }
                },
                "device_types": {
                    "description": "",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/DeviceType"
                    }
                }
            },
            "type": "object"
        },
...
        "Kit": {
            "properties": {
                "id": {
                    "description": "",
                    "type": "integer",
                    "format": "int"
                },
                "name": {
                    "description": "",
                    "type": "integer",
                    "format": "string"
                },
                "created_at": {
                    "description": "",
                    "type": "string",
                    "format": "date"
                },
                "updated_at": {
                    "description": "",
                    "type": "string",
                    "format": "date"
                }
            },
            "type": "object"
        },
...
        "SubType": {
            "properties": {
                "id": {
                    "description": "",
                    "type": "integer",
                    "format": "int"
                },
                "label": {
                    "description": "",
                    "type": "integer",
                    "format": "string"
                },
                "description": {
                    "description": "",
                    "type": "string",
                    "format": "string"
                },
                "created_at": {
                    "description": "",
                    "type": "string",
                    "format": "date"
                },
                "updated_at": {
                    "description": "",
                    "type": "string",
                    "format": "date"
                }
            },
            "type": "object"
        },
        "TokenExpired": {
            "properties": {
                "token": {
                    "description": "",
                    "type": "string",
                    "format": "string"
                },
                "error": {
                    "description": "",
                    "type": "string",
                    "format": "string"
                }
            },
            "type": "object"
        },
...
    },
...
}